bug fixes and improvements

+++ index.ts
@@ -6,8 +6,8 @@ import db from './firestore'
 import { firestore } from 'firebase-admin'
 import { stripIndent } from 'common-tags'
 
-import { commandIDs, token } from './config.json'
-import { check, no } from './emoji.json'
+import { commandIDs, emoji, token } from './config.json'
+const { check, no } = emoji
 
 const rest = new REST().setToken(token)
 
@@ -58,7 +58,7 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 	} else if (interaction.data.id === commandIDs.ping) {
 		await api.interactions.defer(interaction.id, interaction.token)
 		const reply = await api.interactions.getOriginalReply(interaction.application_id, interaction.token)
-		const start = Number(BigInt(interaction.id) >> 22n) + 1420070400000
+		const start = Math.floor(Number(interaction.id) / 2 ** 22) + 1420070400000;
 		const end = new Date(reply.timestamp).getTime()
 		api.interactions.editReply(interaction.application_id, interaction.token, {
 			content: `🏓 Pong! Took **${end - start}**ms.`
@@ -67,7 +67,7 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 	} else if (interaction.data.id === commandIDs.help) {
 		respond({
 			content: stripIndent`
-				Usertags is a version of [Slashtags](<https://github.com/advaith1/slashtags>) for global user tags.
+				[Usertags](<https://github.com/advaith1/usertags>) is a version of [Slashtags](<https://github.com/advaith1/slashtags>) for global user tags.
 				* Use </create:${commandIDs.create}> to create a tag, </edit:${commandIDs.edit}> to edit a tag, and </delete:${commandIDs.delete}> to delete a tag.
 				* Use </tag:${commandIDs.tag}> to send a tag.
 				Created by [advaith](<https://advaith.io>)`
@@ -75,9 +75,9 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 
 	} else if (interaction.data.id === commandIDs.create) {
 		if (option('name')?.length > 100)
-			return respond({ content: no+'name must be 1 to 100 characters' })
+			return respond({ content: no+'name must be 1 to 100 characters', flags: MessageFlags.Ephemeral })
 		if (!option('content') || option('content').length > 2000)
-			return respond({ content: no+'content must be 1 to 2000 characters' })
+			return respond({ content: no+'content must be 1 to 2000 characters', flags: MessageFlags.Ephemeral })
 
 		let error = false
 
@@ -88,7 +88,7 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 			}
 		}, { merge: true }).catch(e => {
 			error = true
-			respond({ content: no+`error: ${e}`, allowed_mentions: { parse: [] } })
+			respond({ content: no+`error: ${e}`, flags: MessageFlags.Ephemeral, allowed_mentions: { parse: [] } })
 		})
 
 		if (!error) {
@@ -97,18 +97,18 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 		}
 
 	} else if (interaction.data.id === commandIDs.edit) {
-		if ((!option('newname') || option('newname') === option('tag')) && !option('content') && option('ephemeral') === null)
-			return respond({ content: no+"you didn't provide any new data!" })
+		if ((!option('newname') || option('newname') === option('tag')) && !option('content') && option('ephemeral') === undefined)
+			return respond({ content: no+"you didn't provide any new data!", flags: MessageFlags.Ephemeral })
 		if (option('newname')?.length > 100)
-			return respond({ content: no+'newname must be 1 to 100 characters' })
+			return respond({ content: no+'newname must be 1 to 100 characters', flags: MessageFlags.Ephemeral })
 		if (option('content')?.length > 2000)
-			return respond({ content: no+'content must be 1 to 2000 characters' })
+			return respond({ content: no+'content must be 1 to 2000 characters', flags: MessageFlags.Ephemeral })
 
 		let error = false
 
 		const doc = await db.collection('users').doc(userID).get()
-		const tag = doc.data()[option('tag')] as Tag
-		if (!tag) return respond({ content: no+'tag not found' })
+		const tag = doc.data()?.[option('tag')] as Tag
+		if (!tag) return respond({ content: no+'tag not found', flags: MessageFlags.Ephemeral })
 
 		let data: object
 		if (option('newname')) {
@@ -131,7 +131,7 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 
 		await db.collection('users').doc(userID).set(data, { merge: true }).catch(e => {
 			error = true
-			respond({ content: no+`error: ${e}`, allowed_mentions: { parse: [] } })
+			respond({ content: no+`error: ${e}`, flags: MessageFlags.Ephemeral, allowed_mentions: { parse: [] } })
 		})
 
 		if (!error) {
@@ -143,11 +143,15 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 	} else if (interaction.data.id === commandIDs.delete) {
 		let error = false
 
+		const doc = await db.collection('users').doc(userID).get()
+		const tag = doc.data()?.[option('tag')] as Tag
+		if (!tag) return respond({ content: no+'tag not found', flags: MessageFlags.Ephemeral })
+
 		await db.collection('users').doc(userID).update({
 			[option('tag')]: firestore.FieldValue.delete()
 		}).catch(e => {
 			error = true
-			respond({ content: no+`error: ${e}`, allowed_mentions: { parse: [] } })
+			respond({ content: no+`error: ${e}`, flags: MessageFlags.Ephemeral, allowed_mentions: { parse: [] } })
 		})
 
 		if (!error) {
@@ -157,8 +161,8 @@ client.on(GatewayDispatchEvents.InteractionCreate, async ({ data: interaction, a
 
 	} else if (interaction.data.id === commandIDs.tag) {
 		const doc = await db.collection('users').doc(userID).get()
-		const tag = doc.data()[option('tag')] as Tag
-		if (!tag) return respond({ content: no+'tag not found' })
+		const tag = doc.data()?.[option('tag')] as Tag
+		if (!tag) return respond({ content: no+'tag not found', flags: MessageFlags.Ephemeral })
 		respond({ content: tag?.content, flags: tag?.ephemeral ? MessageFlags.Ephemeral : undefined, allowed_mentions: { parse: [] } })
 			.catch(console.log)
 	}
